configs: nintendo switch: do not control uart clock
In case BPMP-FW is used with debugging enabled, it manages and enables the uart clock.
In such cases, if uboot re-enables the clock and BPMP-FW prints something, it will hang.

In all other cases, uart clock and uart console is already managed by bootloader.

Effectively this change, allows again the use of BPMP-FW debugging for testing, which functionality is currently forced disabled.